chore: separate firebase client and server sdks
