+++ src/lib/firebase/client.ts
@@ -1,20 +1,5 @@
-import * as admin from 'firebase-admin';
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
-
-if (!admin.apps.length) {
-    admin.initializeApp({
-        credential: admin.credential.cert({
-            projectId: process.env.FIREBASE_PROJECT_ID,
-            privateKey: process.env.FIREBASE_PRIVATE_KEY?.replace(/\\n/g, '\n'),
-            clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
-        }),
-        projectId: process.env.FIREBASE_PROJECT_ID,
-    });
-}
-
-export const db = admin.firestore();
-export const auth = admin.auth();
 
 const firebaseConfig = {
     apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
