+++ src/lib/firebase/server.ts
@@ -1,0 +1,15 @@
+import * as admin from 'firebase-admin';
+
+if (!admin.apps.length) {
+    admin.initializeApp({
+        credential: admin.credential.cert({
+            projectId: process.env.FIREBASE_PROJECT_ID,
+            privateKey: process.env.FIREBASE_PRIVATE_KEY?.replace(/\\n/g, '\n'),
+            clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
+        }),
+        projectId: process.env.FIREBASE_PROJECT_ID,
+    });
+}
+
+export const db = admin.firestore();
+export const auth = admin.auth();
